requirements.txt
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-core
langchain-community
openai==0.28
fastapi
chromadb
unstructured
"unstructured[all]"
coloredlogs